[REL] migration product_pack to 11.0  (#145)
* [FIX] xml files names

* [REL] migration product_pack to 11.0
* [FIX] Update README
* [FIX] remove openerp tags and data tags
* [IMP] change some funtion to improve funcionality
* [FIX] change name to xml and .py files according guidelines
* [FIX] remove api.one decorators

* [FIX] update readme

* [FIX] Remove "encoding: utf-8"

* [FIX] remove api.multi, string not necesary and some other fixes

* [FIX] demo data to separate according Guidelines

* [FIX] fix names in fields

* [FIX] remove executable
