+++ product_pack/README.rst
@@ -1,12 +1,22 @@
-.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
-  :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
-  :alt: License: AGPL-3
+.. |company| replace:: ADHOC SA
+
+.. |company_logo| image:: https://raw.githubusercontent.com/ingadhoc/maintainer-tools/master/resources/adhoc-logo.png
+   :alt: ADHOC SA
+   :target: https://www.adhoc.com.ar
+
+.. |icon| image:: https://raw.githubusercontent.com/ingadhoc/maintainer-tools/master/resources/adhoc-icon.png
+
+.. image:: https://img.shields.io/badge/license-AGPL--3-blue.png
+   :target: https://www.gnu.org/licenses/agpl
+   :alt: License: AGPL-3
 
 ============
 Product Pack
 ============
 
-This module add "Product Pack" funcionality to sale orders. For prouct you can add some other product to sell with them, in the sale order you choose the product and then automaticaly added the products that set in Product selected. 
+This module add "Product Pack" funcionality to sale orders. For product you can add some other product to sell with them, in the sale order you choose the product and then automaticaly added the products that set in Product selected.
+
+TODO: check modifications of this fork https://github.com/mohamedhagag/dvit-odoo/tree/10.0/product_pack
 
 Installation
 ============
@@ -14,7 +24,6 @@ Installation
 To install this module, you need to:
 
 #. Just install this module.
-
 
 Configuration
 =============
@@ -24,8 +33,6 @@ To configure this module, you need to:
 #. Go to product and set "Product Pack", then set the products has to include in the product pack of the product.
 #. Set type of product pack.
 
-
-
 Usage
 =====
 
@@ -33,14 +40,9 @@ To use this module, you need to:
 
 #. Create an sale order and set product that has product pack choosed.
 
-
 .. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
-  :alt: Try me on Runbot
-  :target: https://runbot.adhoc.com.ar/
-
-.. repo_id is available in https://github.com/OCA/maintainer-tools/blob/master/tools/repos_with_ids.txt
-.. branch is "9.0" for example
-
+   :alt: Try me on Runbot
+   :target: http://runbot.adhoc.com.ar/
 
 Bug Tracker
 ===========
@@ -56,19 +58,16 @@ Credits
 Images
 ------
 
-* ADHOC SA: `Icon <http://fotos.subefotos.com/83fed853c1e15a8023b86b2b22d6145bo.png>`_.
+* |company| |icon|
 
 Contributors
 ------------
 
-
 Maintainer
 ----------
 
-.. image:: http://fotos.subefotos.com/83fed853c1e15a8023b86b2b22d6145bo.png
-  :alt: Odoo Community Association
-  :target: https://www.adhoc.com.ar
+|company_logo|
 
-This module is maintained by the ADHOC SA.
+This module is maintained by the |company|.
 
 To contribute to this module, please visit https://www.adhoc.com.ar.
