+++ product_pack/demo/product_pack_line_demo.xml
@@ -1,0 +1,116 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo noupdate="1">
+    <record id="pack_cpu_detailed_components_1" model="product.pack.line">
+        <field name="parent_product_id" ref="product_pack_cpu_detailed_components"/>
+        <field name="product_id" ref="product.product_product_20"/>
+        <field name="quantity" eval="1.0"/>
+        <field name="discount" eval="10.0"/>
+    </record>
+     <record id="pack_cpu_detailed_components_3" model="product.pack.line">
+        <field name="parent_product_id" ref="product_pack_cpu_detailed_components"/>
+        <field name="product_id" ref="product.product_product_16"/>
+        <field name="quantity" eval="1.0"/>
+        <field name="discount" eval="10.0"/>
+    </record>
+    <record id="pack_cpu_detailed_components_4" model="product.pack.line">
+        <field name="parent_product_id" ref="product_pack_cpu_detailed_components"/>
+        <field name="product_id" ref="product.product_product_24"/>
+        <field name="quantity" eval="1.0"/>
+    </record>
+    <record id="pack_cpu_detailed_totaliced_1" model="product.pack.line">
+        <field name="parent_product_id" ref="product_pack_cpu_detailed_totaliced"/>
+        <field name="product_id" ref="product.product_product_20"/>
+        <field name="quantity" eval="1.0"/>
+        <field name="discount" eval="10.0"/>
+    </record>
+    <record id="pack_cpu_detailed_totaliced_3" model="product.pack.line">
+        <field name="parent_product_id" ref="product_pack_cpu_detailed_totaliced"/>
+        <field name="product_id" ref="product.product_product_16"/>
+        <field name="quantity" eval="1.0"/>
+        <field name="discount" eval="10.0"/>
+    </record>
+    <record id="pack_cpu_detailed_totaliced_4" model="product.pack.line">
+        <field name="parent_product_id" ref="product_pack_cpu_detailed_totaliced"/>
+        <field name="product_id" ref="product.product_product_24"/>
+        <field name="quantity" eval="1.0"/>
+    </record>
+     <record id="pack_cpu_detailed_fixed_1" model="product.pack.line">
+        <field name="parent_product_id" ref="product_pack_cpu_detailed_fixed"/>
+        <field name="product_id" ref="product.product_product_20"/>
+        <field name="quantity" eval="1.0"/>
+    </record>
+    <record id="pack_cpu_detailed_fixed_3" model="product.pack.line">
+        <field name="parent_product_id" ref="product_pack_cpu_detailed_fixed"/>
+        <field name="product_id" ref="product.product_product_16"/>
+        <field name="quantity" eval="1.0"/>
+    </record>
+    <record id="pack_cpu_detailed_fixed_4" model="product.pack.line">
+        <field name="parent_product_id" ref="product_pack_cpu_detailed_fixed"/>
+        <field name="product_id" ref="product.product_product_24"/>
+        <field name="quantity" eval="1.0"/>
+    </record>
+    <record id="pack_cpu_none_detailed_assited_1" model="product.pack.line">
+        <field name="parent_product_id" ref="product_pack_cpu_none_detailed_assited"/>
+        <field name="product_id" ref="product.product_product_20"/>
+        <field name="quantity" eval="1.0"/>
+        <field name="discount" eval="10.0"/>
+    </record>
+    <record id="pack_cpu_none_detailed_assited_3" model="product.pack.line">
+        <field name="parent_product_id" ref="product_pack_cpu_none_detailed_assited"/>
+        <field name="product_id" ref="product.product_product_16"/>
+        <field name="quantity" eval="1.0"/>
+        <field name="discount" eval="10.0"/>
+    </record>
+    <record id="pack_cpu_none_detailed_assited_4" model="product.pack.line">
+        <field name="parent_product_id" ref="product_pack_cpu_none_detailed_assited"/>
+        <field name="product_id" ref="product.product_product_24"/>
+        <field name="quantity" eval="1.0"/>
+    </record>
+    <record id="pack_cpu_none_detailed_totaliced_1" model="product.pack.line">
+        <field name="parent_product_id" ref="product_pack_cpu_none_detailed_totaliced"/>
+        <field name="product_id" ref="product.product_product_20"/>
+        <field name="quantity" eval="1.0"/>
+        <field name="discount" eval="10.0"/>
+    </record>
+    <record id="pack_cpu_none_detailed_totaliced_3" model="product.pack.line">
+        <field name="parent_product_id" ref="product_pack_cpu_none_detailed_totaliced"/>
+        <field name="product_id" ref="product.product_product_16"/>
+        <field name="quantity" eval="1.0"/>
+        <field name="discount" eval="10.0"/>
+    </record>
+    <record id="pack_cpu_none_detailed_totaliced_4" model="product.pack.line">
+        <field name="parent_product_id" ref="product_pack_cpu_none_detailed_totaliced"/>
+        <field name="product_id" ref="product.product_product_24"/>
+        <field name="quantity" eval="1.0"/>
+    </record>
+    <record id="pack_pc_detailed_components_1" model="product.pack.line">
+        <field name="parent_product_id" ref="product_pack_pc_detailed_components"/>
+        <field name="product_id" ref="product.product_product_12"/>
+        <field name="quantity" eval="1.0"/>
+    </record>
+    <record id="pack_pc_detailed_components_3" model="product.pack.line">
+        <field name="parent_product_id" ref="product_pack_pc_detailed_components"/>
+        <field name="product_id" ref="product_pack_cpu_detailed_components"/>
+        <field name="quantity" eval="1.0"/>
+    </record>
+     <record id="pack_pc_detailed_totaliced_1" model="product.pack.line">
+        <field name="parent_product_id" ref="product_pack_pc_detailed_totaliced"/>
+        <field name="product_id" ref="product.product_product_12"/>
+        <field name="quantity" eval="1.0"/>
+    </record>
+    <record id="pack_pc_detailed_totaliced_3" model="product.pack.line">
+        <field name="parent_product_id" ref="product_pack_pc_detailed_totaliced"/>
+        <field name="product_id" ref="product_pack_cpu_detailed_totaliced"/>
+        <field name="quantity" eval="1.0"/>
+    </record>
+     <record id="pack_pc_detailed_fixed_1" model="product.pack.line">
+        <field name="parent_product_id" ref="product_pack_pc_detailed_fixed"/>
+        <field name="product_id" ref="product.product_product_12"/>
+        <field name="quantity" eval="1.0"/>
+    </record>
+    <record id="pack_pc_detailed_fixed_3" model="product.pack.line">
+        <field name="parent_product_id" ref="product_pack_pc_detailed_fixed"/>
+        <field name="product_id" ref="product_pack_cpu_detailed_fixed"/>
+        <field name="quantity" eval="1.0"/>
+    </record>
+</odoo>
