+++ product_pack/demo/product_product_demo.xml
@@ -1,0 +1,89 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo noupdate="1">
+    <record id="product_pack_cpu_detailed_components" model="product.product">
+        <field name="name">Pack CPU (Detailed Components Price)</field>
+        <field name="categ_id" ref="product.product_category_4"/>
+        <field name="pack" eval="True"/>
+        <field name="pack_price_type">components_price</field>
+        <field name="standard_price">20.5</field>
+        <field name="list_price">30.75</field>
+        <field name="type">service</field>
+    </record>
+
+
+    <record id="product_pack_cpu_detailed_totaliced" model="product.product">
+        <field name="name">Pack CPU (Detailed Totaliced Price)</field>
+        <field name="categ_id" ref="product.product_category_4"/>
+        <field name="pack" eval="True"/>
+        <field name="pack_price_type">totalice_price</field>
+        <field name="standard_price">20.5</field>
+        <field name="list_price">30.75</field>
+        <field name="type">service</field>
+    </record>
+
+
+    <record id="product_pack_cpu_detailed_fixed" model="product.product">
+        <field name="name">Pack CPU (Detailed Fixed Price)</field>
+        <field name="categ_id" ref="product.product_category_4"/>
+        <field name="pack" eval="True"/>
+        <field name="pack_price_type">fixed_price</field>
+        <field name="standard_price">20.5</field>
+        <field name="list_price">30.75</field>
+        <field name="type">service</field>
+    </record>
+
+
+    <record id="product_pack_cpu_none_detailed_assited" model="product.product">
+        <field name="name">Pack CPU (None Detailed Assisted Price)</field>
+        <field name="categ_id" ref="product.product_category_4"/>
+        <field name="pack" eval="True"/>
+        <field name="pack_price_type">none_detailed_assited_price</field>
+        <field name="standard_price">20.5</field>
+        <field name="list_price">30.75</field>
+        <field name="type">service</field>
+    </record>
+
+        
+    <record id="product_pack_cpu_none_detailed_totaliced" model="product.product">
+        <field name="name">Pack CPU (None Detailed Totaliced Price)</field>
+        <field name="categ_id" ref="product.product_category_4"/>
+        <field name="pack" eval="True"/>
+        <field name="pack_price_type">none_detailed_totaliced_price</field>
+        <field name="standard_price">20.5</field>
+        <field name="list_price">30.75</field>
+        <field name="type">service</field>
+    </record>
+
+        
+    <record id="product_pack_pc_detailed_components" model="product.product">
+        <field name="name">Pack PC (Detailed Components Price)</field>
+        <field name="categ_id" ref="product.product_category_4"/>
+        <field name="pack" eval="True"/>
+        <field name="pack_price_type">components_price</field>
+        <field name="standard_price">20.5</field>
+        <field name="list_price">30.75</field>
+        <field name="type">service</field>
+    </record>
+
+
+    <record id="product_pack_pc_detailed_totaliced" model="product.product">
+        <field name="name">Pack PC (Detailed Totaliced Price)</field>
+        <field name="categ_id" ref="product.product_category_4"/>
+        <field name="pack" eval="True"/>
+        <field name="pack_price_type">totalice_price</field>
+        <field name="standard_price">20.5</field>
+        <field name="list_price">30.75</field>
+        <field name="type">service</field>
+    </record>
+
+       
+    <record id="product_pack_pc_detailed_fixed" model="product.product">
+        <field name="name">Pack PC (Detailed Fixed Price)</field>
+        <field name="categ_id" ref="product.product_category_4"/>
+        <field name="pack" eval="True"/>
+        <field name="pack_price_type">fixed_price</field>
+        <field name="standard_price">20.5</field>
+        <field name="list_price">30.75</field>
+        <field name="type">service</field>
+    </record>
+</odoo>
