+++ product_pack/models/__init__.py
@@ -1,10 +1,11 @@
-# -*- encoding: utf-8 -*-
 ##############################################################################
-# For copyright and license notices, see __manifest__.py file in root directory
+# For copyright and license notices, see __manifest__.py file in module root
+# directory
 ##############################################################################
 
-from . import pack
-from . import product
+from . import product_pack_line
+from . import product_product
+from . import product_template
 from . import sale_order_line_pack_line
 from . import sale_order_line
 from . import sale_order
