+++ product_pack/models/sale_order_line_pack_line.py
@@ -1,6 +1,6 @@
-# -*- encoding: utf-8 -*-
 ##############################################################################
-# For copyright and license notices, see __manifest__.py file in root directory
+# For copyright and license notices, see __manifest__.py file in module root
+# directory
 ##############################################################################
 from odoo import fields, models, api
 import odoo.addons.decimal_precision as dp
@@ -14,6 +14,7 @@ class SaleOrderLinePackLine(models.Model):
         'sale.order.line',
         'Order Line',
         ondelete='cascade',
+        index=True,
         required=True
     )
     product_id = fields.Many2one(
@@ -31,7 +32,7 @@ class SaleOrderLinePackLine(models.Model):
         digits=dp.get_precision('Discount'),
     )
     price_subtotal = fields.Float(
-        compute="_amount_line",
+        compute="_compute_price_subtotal",
         string='Subtotal',
         digits=dp.get_precision('Account')
     )
@@ -41,14 +42,14 @@ class SaleOrderLinePackLine(models.Model):
         required=True
     )
 
-    @api.one
     @api.onchange('product_id')
     def onchange_product_id(self):
-        self.price_unit = self.product_id.lst_price
+        for line in self:
+            line.price_unit = line.product_id.lst_price
 
-    @api.one
     @api.depends('price_unit', 'product_uom_qty')
-    def _amount_line(self):
-        self.price_subtotal = (
-            self.product_uom_qty * self.price_unit *
-            (1 - (self.discount or 0.0) / 100.0))
+    def _compute_price_subtotal(self):
+        for line in self:
+            line.price_subtotal = (
+                line.product_uom_qty * line.price_unit *
+                (1 - (line.discount or 0.0) / 100.0))
