--- product_pack/views/pack_view.xml
@@ -1,72 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-<data>
-
-	<!-- Product view modification -->
-	<record model="ir.ui.view" id="product_template_only_form_view">
-		<field name="name">product.template.pack.form</field>
-		<field name="model">product.template</field>
-		<field name="inherit_id" ref="product.product_template_only_form_view" />
-		<field name="arch" type="xml">
-			<field name="type" position="after">
-				<field name="pack"/>
-				<field name="pack_price_type" attrs="{'invisible':[('pack', '=', False)], 'required':[('pack','=',True)]}" context="{'pack_price_type': pack_price_type}"/>
-			</field>
-			<notebook position="inside">
-				<page string="Pack" attrs="{'invisible': ['|', ('product_variant_count', '>', 1), ('pack', '=', False)]}">
-					<field name="id" invisible="1"/>
-					<p attrs="{'invisible': [('id', '!=', False)]}">You must save first to add pack lines</p>
-					<field name="pack_line_ids" attrs="{'readonly': [('id', '=', False)]}"/>
-				</page>
-			</notebook>
-		</field>
-	</record>
-
-	<record model="ir.ui.view" id="nan_product_pack_form">
-		<field name="name">product.product.pack.form</field>
-		<field name="model">product.product</field>
-		<field name="inherit_id" ref="product.product_normal_form_view" />
-		<field name="arch" type="xml">
-			<field name="type" position="after">
-				<field name="pack"/>
-				<field name="pack_price_type" attrs="{'invisible':[('pack', '=', False)], 'required':[('pack','=',True)]}" context="{'pack_price_type': pack_price_type}"/>
-			</field>
-			<notebook position="inside">
-				<page string="Pack" attrs="{'invisible':[('pack', '=', False)]}">
-					<field name="pack_line_ids"/>
-				</page>
-			</notebook>
-		</field>
-	</record>
-
-	<!-- product.pack.line -->
-	<record model="ir.ui.view" id="product_pack_line_form">
-		<field name="name">product.pack.line.form</field>
-		<field name="model">product.pack.line</field>
-		<field name="type">form</field>
-		<field name="arch" type="xml">
-			<form string="Pack Product Line">
-				<group>
-					<field name="product_id"/>
-					<field name="quantity"/>
-					<field name="discount" groups="sale.group_discount_per_so_line" invisible="context.get('pack_price_type') == 'fixed_price'"/>
-				</group>
-			</form>
-		</field>
-	</record>
-
-	<record model="ir.ui.view" id="product_pack_line_tree">
-		<field name="name">product.pack.line.tree</field>
-		<field name="model">product.pack.line</field>
-		<field name="type">tree</field>
-		<field name="arch" type="xml">
-			<tree string="Pack Product Lines" editable="bottom">
-				<field name="product_id"/>
-				<field name="quantity"/>
-				<field name="discount" groups="sale.group_discount_per_so_line" invisible="context.get('pack_price_type') == 'fixed_price'"/>
-			</tree>
-		</field>
-	</record>
-</data>
-</odoo>
-
