+++ product_pack/views/product_product_views.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <record model="ir.ui.view" id="nan_product_pack_form">
+        <field name="name">product.product.pack.form</field>
+        <field name="model">product.product</field>
+        <field name="inherit_id" ref="product.product_normal_form_view" />
+        <field name="arch" type="xml">
+            <field name="type" position="after">
+                <field name="pack"/>
+                <field name="pack_price_type" attrs="{'invisible':[('pack', '=', False)], 'required':[('pack','=',True)]}" context="{'pack_price_type': pack_price_type}"/>
+            </field>
+            <notebook position="inside">
+                <page string="Pack" attrs="{'invisible':[('pack', '=', False)]}">
+                    <field name="pack_line_ids"/>
+                </page>
+            </notebook>
+        </field>
+    </record>
+</odoo>
