+++ product_pack/views/sale_order_line_pack_line_views.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <!-- sale order line pack line -->
+    <record model="ir.ui.view" id="view_sale_line_pack_line_tree">
+        <field name="name">sale.line.pack.line.tree</field>
+        <field name="model">sale.order.line.pack.line</field>
+        <field name="arch" type="xml">
+            <tree string="Sales Order Line Pack Lines" editable="bottom">
+                <field name="product_id" groups="base.group_user"/>
+                <field name="price_unit"/>
+                <field name="product_uom_qty"/>
+                <field name="discount" groups="sale.group_discount_per_so_line"/>
+                <field name="price_subtotal"/>
+            </tree>
+        </field>
+    </record>
+</odoo>
