+++ product_pack/views/sale_order_line_views.xml
@@ -1,22 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-
-    <!-- sale order line pack line -->
-    <record model="ir.ui.view" id="view_sale_line_pack_line_tree">
-        <field name="name">sale.line.pack.line.tree</field>
-        <field name="model">sale.order.line.pack.line</field>
-        <field name="arch" type="xml">
-            <tree string="Sales Order Line Pack Lines" editable="bottom">
-                <field name="product_id" groups="base.group_user"/>
-                <field name="price_unit"/>
-                <field name="product_uom_qty"/>
-                <field name="discount" groups="sale.group_discount_per_so_line"/>
-                <field name="price_subtotal"/>
-            </tree>
-        </field>
-    </record>
-
+<odoo>
    <record model="ir.ui.view" id="view_order_line_form2">
         <field name="name">view.order.line.button</field>
         <field name="model">sale.order.line</field>
@@ -49,23 +32,4 @@
             </form>
         </field>
     </record>
-
-    <record model="ir.ui.view" id="view_order_stock_form">
-        <field name="name">sale.order.form.inherit</field>
-        <field name="model">sale.order</field>
-        <field name="inherit_id" ref="sale.view_order_form"/>
-        <field name="arch" type="xml">
-            <!-- button to open line form -->
-            <xpath expr="//field[@name='order_line']/tree//field[@name='price_subtotal']" position="after">
-                <field name="pack_type" invisible="1"/>
-                <button name="action_pack_detail" string="Detail" attrs="{'invisible':[('pack_type', '!=', 'none_detailed_assited_price')]}" type="object"/>
-            </xpath> 
-            <!-- no agregamos el boton ya que se edita después desde la tree pero si agregamos el campo porque si no da error al guardar al cargar productos tipo pop up -->
-            <xpath expr="//field[@name='order_line']/form" position="inside">
-                <field name="pack_type" invisible="1"/>
-            </xpath> 
-        </field>
-    </record>
-
-    </data>
 </odoo>
