+++ product_pack/views/sale_order_views.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <record model="ir.ui.view" id="view_order_stock_form">
+        <field name="name">sale.order.form.inherit</field>
+        <field name="model">sale.order</field>
+        <field name="inherit_id" ref="sale.view_order_form"/>
+        <field name="arch" type="xml">
+            <!-- button to open line form -->
+            <xpath expr="//field[@name='order_line']/tree//field[@name='price_subtotal']" position="after">
+                <field name="pack_type" invisible="1"/>
+                <button name="action_pack_detail" string="Detail" attrs="{'invisible':[('pack_type', '!=', 'none_detailed_assited_price')]}" type="object"/>
+            </xpath> 
+            <!-- no agregamos el boton ya que se edita después desde la tree pero si agregamos el campo porque si no da error al guardar al cargar productos tipo pop up -->
+            <xpath expr="//field[@name='order_line']/form" position="inside">
+                <field name="pack_type" invisible="1"/>
+            </xpath> 
+        </field>
+    </record>
+</odoo>
